Moved printing fn to print.rs
Moved `print_dispenser_format` and `print_branch` to new file print.rs

Cleaned commands
