+++ data/src/chem_tree.rs
@@ -2,9 +2,9 @@ use crate::chemicals::{Chemical, Ingredient, Reaction, BASES_MAP};
 
 use crate::search_engine::Maps;
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct ChemTree {
-    root: Box<ChemTreeNode>,
+    pub root: Box<ChemTreeNode>,
 }
 
 impl ChemTree {
@@ -14,87 +14,10 @@ impl ChemTree {
         }
     }
 
-    fn get_compound(&self) -> &Reaction {
+    pub fn get_compound(&self) -> &Reaction {
         match &self.root.chemical {
             Chemical::Compound(c) => c,
             _ => panic!("A non compound was placed at root of tree!"),
-        }
-    }
-
-    pub fn print_dispenser_format(&self) {
-        for node in self.root.get_reagents() {
-            let mut count = 1;
-            for recipe in node {
-                println!("\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\");
-                if node.len() > 1 {
-                    println!("----\t{}\t----", self.root.get_id().to_uppercase());
-                    println!("\tRECIPE #{}\n", count);
-                    count += 1;
-                } else {
-                    println!("----    {}    ----\n", self.root.get_id().to_uppercase());
-                }
-
-                let mut bases = String::new();
-                let mut compounds = String::new();
-                let mut ingredients = String::new();
-
-                for reagent in recipe {
-                    let result = reagent.print_branch(0);
-                    match result.0 {
-                        Chemical::Compound(_compound) => {
-                            compounds = format!("{}\n{}", compounds, result.1.as_str());
-                        }
-                        Chemical::Base(_base) => {
-                            bases.push_str(result.1.as_str());
-                        }
-                        Chemical::Ingredient(_ingredient) => {
-                            ingredients.push_str(result.1.as_str());
-                        }
-                    }
-                }
-
-                match self.get_compound().get_required_temp() {
-                    Some(temp) => {
-                        println!("# Required Temperature #");
-                        println!("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
-                        println!("{} K", temp);
-                        println!("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n");
-                    }
-                    None => {}
-                }
-
-                if !compounds.is_empty() || !ingredients.is_empty() {
-                    println!("# Non-base Reagents #");
-                    println!("+++++++++++++++++++++++++++++++++++++");
-
-                    if !compounds.is_empty() {
-                        println!("_____________________________________");
-                        println!("Compounds");
-                        println!("-------------------------------------");
-                        println!("{}", compounds);
-                    }
-
-                    if !ingredients.is_empty() {
-                        println!("___________________________________");
-                        println!("Ingredients");
-                        println!("-----------------------------------");
-                        println!("{}", ingredients);
-                    }
-                    println!("+++++++++++++++++++++++++++++++++++++\n");
-                }
-
-                if !bases.is_empty() {
-                    println!("_____________________________________");
-                    println!("Base Reagents");
-                    println!("-------------------------------------");
-                    println!("{}", bases);
-                    println!("-------------------------------------");
-                }
-
-                println!(
-                    "////////////////////////////////////////////////////////////////////////\n"
-                );
-            }
         }
     }
 
@@ -147,10 +70,10 @@ impl ChemTree {
     }
 }
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct ChemTreeNode {
-    chemical: Chemical,
-    quantity: f32,
+    pub chemical: Chemical,
+    pub quantity: f32,
     reagents: Box<Option<Vec<Vec<ChemTreeNode>>>>,
 }
 
@@ -163,101 +86,8 @@ impl ChemTreeNode {
         self.reagents = Box::new(Some(branches));
     }
 
-    fn get_reagents(&self) -> &Option<Vec<Vec<ChemTreeNode>>> {
+    pub fn get_reagents(&self) -> &Option<Vec<Vec<ChemTreeNode>>> {
         &self.reagents
-    }
-
-    // probably needs to be broken into seperate functions for each reagent type
-    fn print_branch(&self, layer: i8) -> (&Chemical, String) {
-        let result: (&Chemical, String);
-
-        let mut tab = String::new();
-        let mut c = layer;
-        while c > 0 {
-            tab = format!("\t{}", tab);
-            c -= 1;
-        }
-
-        match &self.chemical {
-            Chemical::Compound(compound) => {
-                let mut branch_strings = Vec::new();
-                for top_branch in self.get_reagents() {
-                    let recipe = &top_branch[0]; // Moved Hardcoded use of 1st Recipe here
-                    for node in recipe {
-                        branch_strings.push(node.print_branch(layer + 1));
-                    }
-                }
-
-                let mut bases = String::new();
-                let mut compounds = String::new();
-                let mut ingredients = String::new();
-
-                for s in branch_strings {
-                    match s.0 {
-                        Chemical::Compound(_compound) => {
-                            compounds = format!("{}\n{}", compounds, s.1.as_str());
-                        }
-                        Chemical::Base(_base) => {
-                            bases.push_str(s.1.as_str());
-                        }
-                        Chemical::Ingredient(_ingredient) => {
-                            ingredients.push_str(s.1.as_str());
-                        }
-                    }
-                }
-
-                let mut branch = String::new();
-                if !bases.is_empty() {
-                    branch = format!("\n{tab}\t{}", bases);
-                }
-                if !ingredients.is_empty() {
-                    branch = format!("{branch}\n{tab}\t{}", ingredients);
-                }
-                if !compounds.is_empty() {
-                    branch = format!("{branch}\n{tab}{}", compounds);
-                }
-
-                let compound_value = format!(
-                    "{tab}[{} {}]",
-                    self.quantity,
-                    compound.get_internal_name().to_ascii_uppercase()
-                );
-
-                let temp_val = compound.get_required_temp();
-
-                let recipe = match temp_val {
-                    Some(temp) => {
-                        format!(
-                            "{} (@{}K)\n{tab}{{\n{}\n{tab}}}\n",
-                            compound_value, temp, branch
-                        )
-                    }
-                    None => {
-                        format!("{}\n{tab}{{\n{}\n{tab}}}\n", compound_value, branch)
-                    }
-                };
-
-                result = (&self.chemical, recipe);
-            }
-            Chemical::Base(base) => {
-                result = (
-                    &self.chemical,
-                    format!(
-                        "({} {}) ",
-                        self.quantity,
-                        base.get_id().to_ascii_uppercase()
-                    ),
-                );
-            }
-            Chemical::Ingredient(ingredient) => {
-                result = (
-                    &self.chemical,
-                    format!("<{} \"{}\"> ", self.quantity, ingredient.get_id()),
-                );
-            }
-        }
-
-        result
     }
 }
